correcciones del review
Se realizaron los arreglos sugeridos en el review del PR
