+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/data/local/sqldb/LastFMLocalStorage.kt
@@ -4,7 +4,7 @@ import android.content.ContentValues
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
-import ayds.lisboa.songinfo.moredetails.domain.Biography.ArtistBiography
+import ayds.lisboa.songinfo.moredetails.domain.entities.Biography.ArtistBiography
 
 interface LastFMLocalStorage {
     fun saveArtist(artist: String, artistBiography: ArtistBiography)
