--- app/src/main/java/ayds/lisboa/songinfo/moredetails/domain/BiographyRepository.kt
@@ -1,8 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.domain
-
-interface BiographyRepository {
-    fun getArtistBiography(artistName: String): Biography
-}
-
-
-
