+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/domain/entities/Biography.kt
@@ -1,4 +1,4 @@
-package ayds.lisboa.songinfo.moredetails.domain
+package ayds.lisboa.songinfo.moredetails.domain.entities
 
 sealed class Biography {
