+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/domain/repository/BiographyRepository.kt
@@ -1,0 +1,10 @@
+package ayds.lisboa.songinfo.moredetails.domain.repository
+
+import ayds.lisboa.songinfo.moredetails.domain.entities.Biography
+
+interface BiographyRepository {
+    fun getArtistBiography(artistName: String): Biography
+}
+
+
+
